+++ platform-operator/controllers/verrazzano/component/certmanager/ocidns/certmanagerocidns.go
@@ -5,9 +5,12 @@ package ocidns
 
 import (
 	"fmt"
+	"github.com/verrazzano/verrazzano/pkg/constants"
 	"github.com/verrazzano/verrazzano/pkg/k8s/ready"
+	vzresource "github.com/verrazzano/verrazzano/pkg/k8s/resource"
 	"github.com/verrazzano/verrazzano/pkg/vzcr"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 )
 
@@ -25,4 +28,25 @@ func isCertManagerOciDNSReady(context spi.ComponentContext) bool {
 	deployments = append(deployments, types.NamespacedName{Name: ocidnsDeploymentName, Namespace: ComponentNamespace})
 	prefix := fmt.Sprintf("Component %s", context.GetComponent())
 	return ready.DeploymentsAreReady(context.Log(), context.Client(), deployments, 1, prefix)
+}
+
+func (c certManagerOciDNSComponent) postUninstall(ctx spi.ComponentContext) error {
+	// Clean up the OCI DNS secret in the clusterResourceNamespace
+	dns := ctx.EffectiveCR().Spec.Components.DNS
+	if dns == nil || dns.OCI == nil {
+		return nil
+	}
+	ociDNS := dns.OCI
+
+	err := vzresource.Resource{
+		Name:      ociDNS.OCIConfigSecret,
+		Namespace: constants.CertManagerNamespace,
+		Client:    ctx.Client(),
+		Object:    &corev1.Secret{},
+		Log:       ctx.Log(),
+	}.Delete()
+	if err != nil {
+		return err
+	}
+	return nil
 }
