+++ platform-operator/controllers/verrazzano/component/certmanager/ocidns/certmanagerocidns_component.go
@@ -22,7 +22,7 @@ import (
 const ComponentName = cmcommon.CertManagerOCIDNSComponentName
 
 // ComponentNamespace is the namespace of the component
-const ComponentNamespace = constants.CertManagerNamespace
+const ComponentNamespace = constants.VerrazzanoSystemNamespace
 
 const componentChartName = "verrazzano-cert-manager-ocidns-webhook"
 
@@ -52,7 +52,7 @@ func NewComponent() spi.Component {
 }
 
 func (c certManagerOciDNSComponent) PreInstall(ctx spi.ComponentContext) error {
-	if err := common.CopyOCIDNSSecret(ctx, ComponentNamespace); err != nil {
+	if err := common.CopyOCIDNSSecret(ctx, constants.CertManagerNamespace); err != nil {
 		return err
 	}
 	return nil
@@ -67,6 +67,10 @@ func (c certManagerOciDNSComponent) IsEnabled(effectiveCR runtime.Object) bool {
 		return false
 	}
 	return vzcr.IsOCIDNSEnabled(effectiveCR)
+}
+
+func (c certManagerOciDNSComponent) PostUninstall(ctx spi.ComponentContext) error {
+	return c.postUninstall(ctx)
 }
 
 // IsReady component check
