+++ platform-operator/helm_config/charts/verrazzano-cert-manager-ocidns-webhook/templates/rbac.yaml
@@ -98,15 +98,15 @@ apiVersion: rbac.authorization.k8s.io/v1
 kind: Role
 metadata:
   name: {{ include "cert-manager-webhook-oci.fullname" . }}:secret-reader
-  namespace: {{ .Release.Namespace }}
+  namespace: {{ .Values.certManager.clusterResourceNamespace }}
 rules:
   - apiGroups:
     - ""
     resources:
     - "secrets"
-  {{- with .Values.ociProfileSecretNames }}
+    {{- with .Values.ociProfileSecretNames }}
     resourceNames:
-{{ toYaml . | indent 4 }}
+    {{ toYaml . | indent 4 }}
     {{- end }}
     verbs:
     - "get"
@@ -116,7 +116,7 @@ apiVersion: rbac.authorization.k8s.io/v1
 kind: RoleBinding
 metadata:
   name: {{ include "cert-manager-webhook-oci.fullname" . }}:secret-reader
-  namespace: {{ .Release.Namespace }}
+  namespace: {{ .Values.certManager.clusterResourceNamespace }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: Role
